#486: Highlighting of composed variables within strings. (#701)
Not all of the configured highlighting was applied inside of strings. Now string highlighting does not override the highlighting of contained elements. The Kotlin plugin is handling this in the same way.

Now more highlighting is done via token highlighting and less by the BashAnnotator. This should also speed up highlighting a bit.