#458: Fix parsing of extglob syntax. (#722)
* #458: Fix parsing of extglob syntax.
Extended patterns are now accepted by the lexer. Full blown
parsing of patterns is out-of-scope for now and could be done
in a follow-up issue.
The bang token "!" is now lexed as a word. Before a command it has special
meaning but as a command argument it's a glob pattern. We're now
checking the token text instead of using a custom token type during parsing.
This adds test cases, the fix and a changelog entry.

* #118: Add extglob support for case patterns

* #118 Support extglob patterns in case and parameter expansions.
Pattern lexing is now more strict, an empty pattern like @() now isn't a valid pattern to make this work with the existing parameter expansion pattern tokens.

* element.getNode() could be null

* fix redundant code: move private PipelineParsing.isWord to ParserUtil
and use it in WordParsing and HistoryExpansionParsingFunction

* move test118() from BashLexerTest to new class BashSdkLexerTest