Remove optional dependency on JetBrains Shell. (#752)
* Remove optional dependency on JetBrains Shell.
If we depend on it our completion tests fail (among others), as
the Shell plugin is also contribution completions and more.
Fixing this in gradle is non trivial as we'd had to extract the sources referencing
shell into a new source dir and modify the build only when tests are run.
We're now referencing the ShFileType by name and not by reference to
avoid the dependency on the plugin. This was the only reference to
code of the Shell plugin.

* Reference the file type by simple class name to no depend on values
which may change in the future.